+++ Items/Weapons/Thrown/Hyua.cs
@@ -1,13 +1,6 @@
 ﻿
 using Microsoft.Xna.Framework;
 using Stellamod.Helpers;
-using Stellamod.Items.Harvesting;
-using Stellamod.Items.Materials;
-using Stellamod.Items.Ores;
-using Stellamod.Particles;
-using Stellamod.Projectiles;
-
-using Stellamod.Projectiles.Slashers.Hearstspire;
 using Stellamod.Projectiles.Slashers.Hyua;
 using System.Collections.Generic;
 using Terraria;
@@ -91,11 +84,11 @@ namespace Stellamod.Items.Weapons.Thrown
             int dir = AttackCounter;
             if (player.direction == 1)
             {
-                player.GetModPlayer<CorrectSwing>().SwingChange = (int)AttackCounter;
+                player.GetModPlayer<CorrectSwing>().SwingChange = AttackCounter;
             }
             else
             {
-                player.GetModPlayer<CorrectSwing>().SwingChange = (int)AttackCounter * -1;
+                player.GetModPlayer<CorrectSwing>().SwingChange = AttackCounter * -1;
 
             }
             AttackCounter = -AttackCounter;
