+++ Projectiles/SkullShotKaboom.cs
@@ -6,7 +6,7 @@ using Terraria.ModLoader;
 
 namespace Stellamod.Projectiles
 {
-	public class SkullShotKaboom : ModProjectile
+    public class SkullShotKaboom : ModProjectile
 	{
 		public override void SetStaticDefaults()
 		{
@@ -33,7 +33,7 @@ namespace Stellamod.Projectiles
 
 		public override Color? GetAlpha(Color lightColor)
 		{
-			return new Color(100, 38, 22, 0) * (1f - (float)Projectile.alpha / 50f);
+			return new Color(100, 38, 22, 0) * (1f - Projectile.alpha / 50f);
 		}
 		public override void AI()
         {
